dsl-agent-project/requirements.txt
openai==1.3.0 
pyyaml==6.0.1 
click==8.1.7 
python-dotenv==1.0.0 
pytest==7.4.0 
dashscope==1.20.0  # 通义千问SDK